cargo: Split the project in bin and lib
The project will have a binary and a library.

The binary (vkmsctl) will be the CLI tool to manipulate VKMS devices.

The library will contain reusable code to be used from other projects.

Split the project in these two targets and make sure dependencies are
not used in the library if not required.